Fixed error in uploading author imge

+++ app/Http/Controllers/AuthorsController.php
@@ -91,8 +91,10 @@ class AuthorsController extends Controller
             //Moving image to images folder and saving in database
             $img_path->move('images', $img_name);
             //deleting old file
-            if(file_exists(public_path().'/images/'.$user->img_path)){
-                unlink(public_path().'/images/'.$user->img_path);
+            if($user->img_path){
+                if(file_exists(public_path().'/images/'.$user->img_path)){
+                    unlink(public_path().'/images/'.$user->img_path);
+                }
             }
             $user->img_path = $img_name;
         }
